Fixes rank positioning issues
Ensures that we only ignore ‘duplicate’ listings if it’s the website
we’re looking for as it would return +1 position further down the
results.

There were many cases where the search results included different pages
consecutively from the same domain. This duplicate check caused the
rank to be off by 1 for each of those cases.

Perhaps there should be a configuration option for this as this may be
my personal preference?