Exclude on-page anchor targets as new URLs
Don’t just check for the existence of # in URLs, also look for anything
that links to the same page (#navigation or #button, ect). My only
concern with this is if the link is the *only* link to a page that has
on-page anchor targets, the audit will then likely exclude the page
entirely.

May need to consider comparing against all existing links to see if it
exists already?